Class Challenge with Super and Sub Classes
As a reminder, a subclass is the class that inherits the structure and behavior of another class (called a superclass). Subclasses are also called inherited classes or derived classes. Superclasses are also called base classes. In this diagram, the superclass (base class) is “SHAPE,” and the subclasses (inherited classes/derived classes) are “Rectangle,” “Triangle” and “Circle”: